Debugged Animation Issue
Updated an issue with the animation of the dice when the roll button has been clicked. The Animation continued to run while the information was passed to the answer check function. Added a Settimeout function in the answer check so it will not display the roll until after the animation is completed.