Update app.js
Corrected Issue with the Dice Roll as I erroneously added the setTimeout before the Dice Roll Function was called. Added the Winner Class to the code to add to the appropriate player once the game is completed. Added additional lines to the newGame function to remove winner class from both players once it is called.